+++ birdy/client/base.py
@@ -9,6 +9,7 @@ import owslib
 import requests
 import requests.auth
 from boltons.funcutils import FunctionBuilder
+import owslib
 from owslib.util import ServiceException
 from owslib.wps import (
     ASYNC,
@@ -17,6 +18,8 @@ from owslib.wps import (
     ComplexData,
     WebProcessingService,
 )
+from warnings import warn
+import packaging
 
 from birdy.client import notebook, utils
 from birdy.client.outputs import WPSResult
@@ -46,13 +49,13 @@ class WPSClient(object):
         auth=None,
         verify=True,
         cert=None,
-        verbose=False,
         progress=False,
         version=WPS_DEFAULT_VERSION,
         caps_xml=None,
         desc_xml=None,
         language=None,
         lineage=False,
+        **kwds
     ):
         """Initialize WPSClient.
 
@@ -78,7 +81,7 @@ class WPSClient(object):
         cert: str
           passed to :class:`owslib.wps.WebProcessingService`
         verbose: str
-          passed to :class:`owslib.wps.WebProcessingService`
+          Deprecated. passed to :class:`owslib.wps.WebProcessingService` for owslib < 0.29
         progress: bool
           If True, enable interactive user mode.
         version: str
@@ -117,17 +120,26 @@ class WPSClient(object):
             auth_headers = ["Authorization", "Proxy-Authorization", "Cookie"]
             headers.update({h: r.headers[h] for h in auth_headers if h in r.headers})
 
+        if "verbose" in kwds:
+            if packaging.version.parse(owslib.__version__) >= packaging.version.parse("0.29.0"):
+                kwds.pop("verbose")
+            warn(
+                "The 'verbose' keyword is deprecated and will be removed in a future version. Starting with owslib "
+                "0.29.0, debugging information is logged instead of printed.",
+                DeprecationWarning,
+            )
+
         self._wps = WebProcessingService(
             url,
             version=version,
             username=username,
             password=password,
-            verbose=verbose,
             headers=headers,
             verify=verify,
             cert=cert,
             skip_caps=True,
             language=language,
+            **kwds
         )
 
         try: